separate context into distinct directory
